Provide more meaningful labels in cutlist
we pass down the name of further up items and then when we finally hit
an actual body we make some vague heuristic about how to incorporate
that.

this is because the labels of the actual bodies are often not very
useful and rely on the context of the assembly they reside in.